backend/auth-service/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Configuracion
pydantic-settings==2.2.1

# Base de datos
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1

# JWT
python-jose[cryptography]==3.3.0

# Encriptacion y seguridad
passlib[bcrypt]==1.7.4
cryptography==42.0.7

# TOTP y QR
pyotp==2.9.0
qrcode[pil]==7.4.2

# HTTP cliente (comunicacion entre servicios)
httpx==0.27.0

# Utilidades
python-multipart==0.0.9